requirements.txt
fastapi
gunicorn
uvicorn
requests
pydantic
openpyxl
python-dateutil
